requirements.txt
langchain==0.1.20
langchain-community==0.0.35
langchain-chroma==0.1.1
langchain-groq==0.1.2
langchain-huggingface==0.0.14
langgraph==0.0.53
chromadb==0.4.24
sentence-transformers==2.2.2
groq==0.4.3
pymupdf==1.23.8
pypdf==3.17.1
pulp==2.7.0
scipy==1.11.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.28.1
